+++ R/DeepPatientLevelPrediction.R
@@ -42,7 +42,9 @@
 torch <- NULL
 
 .onLoad <- function(libname, pkgname) {
-  reticulate::py_require(c("polars", "connectorx", "duckdb", "pyarrow", "torch", "tqdm", 
-                           "pynvml"))
+  reticulate::py_require(c(
+    "polars>=1.31.0", "pyarrow", "duckdb", "torch<=2.6.0", "tqdm",
+    "pynvml"
+  ))
   torch <<- reticulate::import("torch", delay_load = TRUE)
 }